Move product data from temp to Postgresql

+++ frontend/src/components/CartProduct.js
@@ -1,23 +1,40 @@
 import { Button } from "react-bootstrap";
 import { CartContext } from "../CartContext";
-import { useContext } from "react";
+import { useContext, useState, useEffect } from "react";
 import { getProductData } from "../productsStore";
 
 function CartProduct(props) {
-  const cart = useContext(CartContext);
+
+  const [productData, setProductData] = useState();
   const id = props.id;
   const quantity = props.quantity;
-  const productData = getProductData(id);
+  useEffect(() => {
+    async function loadData() {
+      const productData = await getProductData(id);
+      setProductData(productData);
+    }
+    loadData();
 
-  return(
+  }, [])
+  const cart = useContext(CartContext);
+  if (!productData) {
+    return (
+      <div>Loading</div>
+    )
+  }
+  // console.log("productData.title from CartPRoduct: ", productData);
+
+
+
+  return (
 
     <>
       <h3>{productData.title}</h3>
       <p>{quantity} total</p>
-      <p>${ (quantity * productData.price).toFixed(2)}</p>
-      <Button size="sm" 
-      onClick={ () => cart.deleteFromCart(id)}>Remove</Button>
-    
+      <p>${(quantity * productData.price).toFixed(2)}</p>
+      <Button size="sm"
+        onClick={() => cart.deleteFromCart(id)}>Remove</Button>
+
     </>
 
   )
